# Conflicts:
#	.gitignore
#	FrontEnd/Components/MatchCard.razor.cs
#	FrontEnd/Components/Pages/Match.razor
#	FrontEnd/FrontEnd.csproj

+++ DTO/Data/Entities/PlayerEntity.cs
@@ -7,7 +7,7 @@ public class PlayerEntity
     [Key]
     public int Id { get; set; }
     public required string Name { get; set; }
-    public int Rating { get; set; } = 1200;
+    public double Rating { get; set; } = 1200;
     
     public virtual ICollection<PlayerMatchEntity> PlayerMatches { get; set; }
 }
